Extract the member access chain walking from the analyzer
Following an append through .EventLog, .Transactional and ?. is syntax
reading with no Chronicle in it, and it had grown to more than a third of
the analyzer. Give it its own type and name, so the analyzer file states
the rule and the new one states how a chain is read.

Pure extraction — no behavior change.
